app/requirements.txt
pytest
requests
pandas
python-dotenv
psycopg2